refactored sockets outside of server.js, various bug fixes
